Created some empty dirs for things that will eventually go there

+++ c_code/biosort.cpp
@@ -1,6 +1,6 @@
 #include "biosort.h"
 
-extern int count;
+extern long long count;
 int * a;
 int i;
 static int random_index = 0;
